Move cache errors to top level cache package without version
