+++ api/internal/cache/BUILD.bazel
@@ -1,8 +1,8 @@
 load("@rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "pkg",
+    name = "cache",
     srcs = ["cache_errors.go"],
-    importpath = "github.com/fjarm/fjarm/api/internal/cache/v1/pkg",
+    importpath = "github.com/fjarm/fjarm/api/internal/cache",
     visibility = ["//api:__subpackages__"],
 )
