+++ api/internal/cache/cache_errors.go
@@ -1,4 +1,4 @@
-package pkg
+package cache
 
 import (
 	"fmt"
