+++ api/internal/cache/v1/pkg/remote/BUILD.bazel
@@ -9,7 +9,7 @@ go_library(
     importpath = "github.com/fjarm/fjarm/api/internal/cache/v1/pkg/remote",
     visibility = ["//api:__subpackages__"],
     deps = [
-        "//api/internal/cache/v1/pkg",
+        "//api/internal/cache",
         "//api/internal/logkeys",
         "@com_github_redis_rueidis//:rueidis",
     ],
@@ -20,7 +20,7 @@ go_test(
     srcs = ["redis_cache_test.go"],
     embed = [":remote"],
     deps = [
-        "//api/internal/cache/v1/pkg",
+        "//api/internal/cache",
         "//api/internal/logkeys",
         "//api/internal/redis/v1/pkg/client",
         "@com_github_redis_rueidis//:rueidis",
