+++ api/internal/cache/v1/pkg/remote/redis_cache.go
@@ -3,7 +3,7 @@ package remote
 import (
 	"context"
 	"fmt"
-	cachev1 "github.com/fjarm/fjarm/api/internal/cache/v1/pkg"
+	cachev1 "github.com/fjarm/fjarm/api/internal/cache"
 	"github.com/fjarm/fjarm/api/internal/logkeys"
 	"github.com/redis/rueidis"
 	"log/slog"
